* Address.cs: change model
* Discount.cs:
* OrderStatus.cs:
* DiscountType.cs:
* Shopia.Domain.csproj:
* Order.cs:
* PaymentStatus.cs:
* DeliveryTimeTable.cs:

+++ Shopia.Domain/Enum/PaymentStatus.cs
@@ -2,7 +2,7 @@
 
 namespace Shopia.Domain
 {
-    public enum PaymentStatus : byte
+    public enum PaymentStatus : int
     {
         Failed = -1,
         Insert = 0,
